[*] tensor/onnx: lift ONNX-inlined logic into tensor methods with builder APIs and input validation
Move operation logic that was duplicated or inlined in ONNX registry
functions into proper Tensor methods with bon builder patterns, making
them reusable outside the ONNX frontend.

Tensor API additions:
- indexing: gather_nd, scatter_nd, tensor_scatter, normalize_negative_indices
- shape_ops: slice_with (Python-style start/end/step/axes)
- lib: cumsum_with, cumprod_with (exclusive + reverse)
- matmul: gemm (alpha * A @ B + beta * C)
- activation: gelu_exact
- nn/mod: depth_to_space, space_to_depth, lrn, nll_loss, dropout, lp_pool
  converted to builder API with snafu::ensure! input validation
- nn/rnn: rnn, gru, lstm converted to builder API with validation
- nn/norm: group_norm converted to builder API; add lp_normalize,
  mean_variance_normalize
- nn/grid_sample: affine_grid, grid_sample converted to builder API
- nn/pool: col2im divisibility check

Input validation (28 ensure! calls across 13 methods):
- 4 new error variants: NdimExact, NdimMinimum, Divisibility, ParamRange
- Replace assert!/assert_eq! panics with proper error returns
- 7 validation tests in test/unit/nn.rs

ONNX registry simplification (-611 lines):
- registry/indexing: replace inline GatherND/ScatterND/TensorScatter/
  CumSum/CumProd with one-liner tensor method calls
- registry/nn: replace inline Gemm/LRN/DepthToSpace/SpaceToDepth/
  LpNorm/MeanVarianceNorm/AffineGrid/GridSample/NllLoss/Dropout/
  GroupNorm with builder calls; remove exact_gelu and dtype_min_positive
  helpers
- registry/shape: replace 100-line inline Slice with slice_with()
- registry/mod: use builder APIs for RNN/GRU/LSTM/Gather/Gelu calls

Supporting changes:
- dtype: add ScalarDType::min_positive() (replaces onnx dtype_min_positive)
- reduce: remove redundant `as usize` cast

+++ tensor/src/reduce.rs
@@ -654,7 +654,7 @@ fn reduce_internal(
         let out_shape: Vec<usize> = shape
             .iter()
             .enumerate()
-            .map(|(i, d)| if resolved_axes.contains(&i) { 1 } else { d.as_const().unwrap_or(1) as usize })
+            .map(|(i, d)| if resolved_axes.contains(&i) { 1 } else { d.as_const().unwrap_or(1) })
             .collect();
 
         let identity = reduction_identity(op, &acc_dtype);
